refactor(codezap): 패키지 이름 변경(thumbnail_snippet에서 언더바 제거)
