+++ backend/src/main/java/codezap/thumbnailsnippet/domain/ThumbnailSnippet.java
@@ -1,4 +1,4 @@
-package codezap.thumbnail_snippet.domain;
+package codezap.thumbnailsnippet.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
