+++ backend/src/main/java/codezap/thumbnailsnippet/repository/ThumbnailSnippetRepository.java
@@ -1,8 +1,8 @@
-package codezap.thumbnail_snippet.repository;
+package codezap.thumbnailsnippet.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import codezap.thumbnail_snippet.domain.ThumbnailSnippet;
+import codezap.thumbnailsnippet.domain.ThumbnailSnippet;
 
 public interface ThumbnailSnippetRepository extends JpaRepository<ThumbnailSnippet, Long> {
 }
